requirements.txt
tweepy
kafka-python
transformers
torch
sklearn
streamlit
plotly
pandas
